rankintimeline
Created the John Rankin timeline and embedded in an html file. Suggestion: please note that these timelines will most likely NOT fit on the main webpages as they are designed to be used in relation to vertical scrolling sites. Therefore, they will need to be a pop-up that appears in a SET sized window with a designated "close" button underneath the timeline.